Removes link to reply on comment when not on a story page.
The reply link does not work on any page that isn't a story page because
it replies on the fact that a comment_form div exists on a page. That
form is the base comment field present on a story page. Therefore,
removing the link when the story is displayed next to the comment
selects for that case easily.

Closes #36 by virtue of removing the link.