mail_new_activity: QP-encode subject
There's no reason to force the subject to be ASCII and mangle many UTF-8
titles when we can just use QP-encoding in the subject like we do in the
body.

This is especially noticable in the IRC channel since the bot mockturtle
uses the mailing list emails to get new stories.

Signed-off-by: Johannes Löthberg <johannes@kyriasis.com>