Add script to ease setup for contributors
Some of the steps in the README are not necessary if all you want to do
is contribute back to Lobsters, rather than run your own instance.
Additionally, the steps that are necessary can be scripted without too
much hassle. `bin/setup` will produce a runnable local version of the
application (without Sphinx).

I also updated the README to break up the instructions for running a
local version of Lobsters and configuring your own version for
deployment.