Update views with email entry to use email field
The views that have been updated:

- settings/index
- invitations/build
- login/forgot_password
- invitation partial
- login/index

In each of these views, there is an email entry input.  This change
sets all of the input types to `email` instead of `text`.  This change
is important because without it, mobile browsers (especially on iOS)
will try to spell check the field which doesn't make sense for an email
address.