add a workaround to stop getting exception emails about IpSpoofAttackErrors
Rails' silly IP spoofing exception triggers when evaluating
request.remote_ip, but because this isn't done until the views are
processing, it comes through as a ActionView::Template::Error, which
isn't ignored by exception notifier.  Eagerly evaluate the IP in the
controller so this raises a proper IpSpoofAttackError, and is
ignored.