search: add tag scoped search
When searching for a story a user can include a single tag:<tag_name>
element which will limit the scope of the search to that tag name.

This works similarly to the domain:<domain_name> feature which
limits the scope of the search to stories from that domain.

Examples:
"tag:python django" as a search query will search for stories
tagged with the python tag which reference the word "django".

"tag:python" will return all of the stories tagged with python
in the search results - which is essentially the /t/python page
but you can now order by Newest or Points as well.

Technical details:
* Added query parsing at the beginning of the search
to set a tag_scope string.

* When a tag_scope is set the relation gets another where clause
for only the tags that match exactly that scope.

* Modified the relevance queries to use literal SQL for the select
statements and the ranked ordering. See: title_match_sql, etc.
When the query was limited to tags ActiveRecord was re-writing
it by substituting the ORDER clauses into the columns for selecting
and the 'AS rel_title' was getting dropped. This caused the query
to fail.