Add some microformat tags to stories.
This commit is part of the webmentions support. These tags help the website
receiving the webmention parsing the mention. Only a minimal amount of
informations have been tagged: I did not wanted to disrupt a lot of views.

We still could extend the h-entry by including the stories comment in them. It
would however spread this whole tagging across several views instead of being
contained in a single one.