fix exception when no user is logged in
Odd production exceptions this morning from story_params that @user was nil
when it checked @user.is_moderator. I think a user was scripting the story
submission form or cleared cookies after opening it. I'm deliberately not
adding the obvious nil check to story_params because non-users shouldn't be
submitting story data; I want the exception that the before_action is missing.