encode fetched content to utf8
Net::HTTP is basically returning raw bytes and letting the user figure out if
the Content-Type header is correct, or if the response is HTML and an included
charset meta tag can be trusted. Rather than trying to get smarter and figure
out the encoding of the page, we're force-encoding to utf8 (which the vast
majority of pages are already) in a way that hopefully won't raise any
exceptions. We don't need perfect data here; good data is plenty.


Fix #590