Remove jQuery from Lobsters.saveStory()
Polyfilled Element.matches & Element.closest so I can use this elsewhere.

Tested by going to both the homepage and a story page and clicking the upvote icons on the story and a comment. Checked that a.) classes are set correctly (styling) and b.) correct network requests are sent.

I don't have an account on lobste.rs so it was difficult to check wether the requests were 1-for-1 however it looks like u{p,n}nvotes were persisted correctly so I assume it's all good.