Send post requests with proper CSRF tokens set
In a previous commit I mentioned "I have no idea how this works" - well now I do. It wasn't and I was testing the wrong feature. I pulled in the `rails-ujs` bundle which is a bundle-size regression for the moment but I checked out the source code and it's pretty small. It doesn't offer a lot of functionality other than wiring up all the default "rails-isms" but seems like the "Rails way" of doing things so is probably OK for the moment. When jQuery is removed we'll still be at a smaller bundle so it seems like a small, temporary inconvenience.

Reorganized application layout to give `application.js` access to the CSRF token in module scope.

It's been a while since I've Rails'd 😄