Convert _showDownvoteWhyAt to not use jQuery
The most tricky thing here was removing the jQuery UI method `position` which is used to overlay popovers etc. I kind of hacked it very simply by not having any collision logic and just setting the position blindly. It's not perfect, but it's a lot less code and it's a mod only feature so it's probably (definitely) "good enough".

Again, tested by clicking around all this stuff a bunch.