Reply to a message as the recipient
I was assuming that we could just take the conversation
author/recipient and apply that to the same fields in the message
object, but that's not true if the recipient is the one composing the
message!

This ensures that the author of the _message_ is the one writing it,
which is not necessarily the author of the conversation.