make Homepage spec more reliable
I'm not sure why this was failing, but for whatever reason I would run
the whole test suite and it would fail once in every ~2-4 runs. It kept
looking for a story title that wasn't on the homepage. I switched the
test to look for a title that I had created manually and that seems to
have made it much more reliable. After 20+ whole-suite runs I got no
failures.