Delete conversations once both sides have marked them as deleted
When a user "deletes" a conversation, it sets the deleted_by_author_at
or deleted_by_recipient column to the current timestamp. Every time the
message is updated, If both are set to a timestamp, we can check that
there have been no messages since the both users have marked the
conversation deleted and actually destroy the record.