Better colours for error boxes
The boxes are hard to read due to the black text on dark red background
colour (contrast ratio: 4.04).

This replaces the background colour with a light red one (ratio: 14.97)
for better readability.

Alternatively, could also muck about with the text colour, but this
looked better IMHO and it was easier to get a good contract ratio
(changing the text to #fff is more readable, but contrast ratio is only
5.19).