Basic conversation model
This sets up a conversation model which has two users, author and
recipient, and also adds a conversation_id to each message. A user's
conversations are fetched with Conversation.involving(user) which take
a user and uses an OR query to find conversations where that user is a
recipient or author.