Convert Convo/Message Creators to Forms
I ran into a bug where message validation errors would not bubble up to
the conversation form. This led me to realize that my abstraction of
creator objects works great if everything goes well, but poorly if they
don't. How do you expose errors outside of the creator objects into
forms? How do you enforce that the conversation is saved only of the
message is valid without a mess of conditionals, especially when the
message needs the conversation?

I chose to switch to form objects that can expose validations and error
messages to a form in a view. This was a bunch of work, but it does
accomplish the same goals as the creator but handles validations and
error messages also.