Turn SSL certificate verification on for critical services (closes #434)
The `Sponge` class used to skip SSL certificate verification entirely
for all HTTPS calls, which makes Lobsters vulnerable to some MITM
attacks that are easily exploited on private or corporate LANs. We
still want to skip certificate verification when there's no specific
security risk, but we do want to validate certificates when
communicating with well-known APIs, such as Archive.is, GitHub,
Gravatar and PushHover. Note that some APIs are accessed outside of
Sponge, such as Twitter's one.

Turning this verification on will flat out break requests to those
websites in case their certificate expired or is incorrect. This may
trigger a 500 for the users of Lobsters, which is what we want. Note
that for Gravatar the 500 would be limited to avatar serving, and
wouldn't impact the main page of Lobsters.