Refactor: move ReadRibbon objects management inside HiddenStory class
Actions on a HiddenStory should always trigger the ReadRibbon
appropriate behavior, we better have them collocated so hiding/unhiding
a story can be done from CLI or somewhere else in the future.

This has the side-effect of changing HiddenStory methods return value,
but it's unused and unlikely to be imho.