PDF parsing support for fetch title
I submit a lot of PDFs, and currently, I have to manually scrape the
title of the PDF. This involves (often troublesome due to poor OCR or
layout by the producing software) copying it from the document, or its
metadata. I developed this feature to automate it, at least for the PDFs
nice enough to have proper metadata filed.

This uses the pdf-reader gem, instead of doing the heavy lifting. It
also refactors the story model's attribute handling to account for items
that aren't HTML. (We can theoretically also parse image/audio/video
metadata too.)

With more refactoring, we can add other information too, like tags or
checking the self-post checkbox, but that's for another day.

This is the first Ruby code I've written, so this might not fly to style
guidelines or whatever.