Add accessor for tag suggestion regex.
Note that right now I can't think of a good way of validating it,
as the difference between what Ruby considers a valid regex and
what Javascript considers a valid regex is a little different.

Similarly, I'm not validating size because some of these regexes
are going to eventually be things like lists of synonyms.

These aren't hard and fast, but it's a starting point.