Fixes merge conflicts
It turns out that I had an old fork of the lobsters code and that's
where I started from without fetching the upstream code so my PR didn't
apply.  This rebases against upstream/master to fix the merge commits.

One of the changes that rebasing showed was that there was a switch to
using the FactoryBot gem for creating test data.  This also updates the
specs that I wrote to use FactoryBot for creating models.