Updates hats index view
This checks that a user is logged in before showing the hat's email on
the hat index page.  This change is needed because a malicious anonymous
user scraped the page and send a user spam based on the email found on
it.

This adds a few new tests around this functionality and handles the
sanitization of the link field in the model.

Issue: #854