revert dark mode (#823)
Firefox support is half-baked. There's no user or devtool UI to toggle between
states, inspecting an element always shows the style for ::selection instead of
the element, and it lists the name of a color variable with no way to see the
value of the variable or where it is set. Punting until it's debuggable.