Only deliver reply notifications on reply+mention
When a reply is made to a parent comment that mentions the parent user,
the parent gets notifications for both the reply and the mention. This
commit makes it so that the reply notification is prioritized, and
mention notifications are only sent when the user didn't directly
reply to the comment.

Fixes #876.