Wrap Story and Comment creation inside of a transcation.
The new reconciliation process between an object's Vote and its
memoized count requires that MariaDB can see the records in the Vote
table. Currently, we are seeing the vote be created, then the
memoization for new comments and stories to zero.

We now force this behavior to take place inside of a transaction.
This should keep the code working the way we expect.