Remember collapsed threads
This change remembers collapsed threads, similar to HN. Unlike HN, it
doesn't store this on the server but in the browser's localStorage. This
is done for both logged in and logged out users.

Alternatively, we could use sessionStorage so it's less persistent. Or
add a "remember collapsed threads" setting with the options
"persistent", "session", and "none". I'm not sure if there are any
people who would consider this feature undesirable?

Tested locally in Firefox and Chromium and seems to work well.