Disable "Post" button after click
This disables the "Post" button after clicking so you can't accidentally
submit it twice.

I also added this to the "reply" button; this loads the HTML from the
server and I've had a few cases where this took a while, and I wasn't
sure if I had clicked it correct so I clicked again and ended up with
two reply fields. There may be other places where this could be added as
well.

It also adds a small loading animation for a bit of feedback.

I also added .fail() for the comment post, since there was no error
before at all.

Partial fix for #967, but it won't cover no-js users.