Reverse the order of messages
Instead of sorting them from old to new this reverses the order from
new to old.

Keeping older messages around is useful for archive purposes, but having
new messages at the bottom is kind of awkward IMO: I need to scroll to
the bottom of the page first.

This is also consistent with how other parts of the site work (i.e.
stories, which also have newer things on top).

There could be a setting for this if some people are attached to the old
sorting.