Remove unused properties in flash banner CSS
The banners' background image gradients have a syntax error, so the background
color is used instead. The second set of border colors, which looks better with
the flat background color, overrides the first set, which looks better with a
corrected gradient. The second text shadow overrides the first as well.

To use gradients, we could replace "top" with "to bottom":
https://developer.mozilla.org/en-US/docs/Web/CSS/gradient/linear-gradient()

But browsers discard these properties, so we have gotten along without them. The
flat appearance matches the site anyway. Therefore this change just removes what
the browser discards.