Use MariaDB instead of MySQL when running tests in GitHub Actions
This starts a Docker container running MariaDB instead of using the MySQL
database that GitHub Actions comes with. This means we can configure which
specific version of MariaDB to run.

I opted for MariaDB 10.5 as that was my best guess as to what version was
running in the deployed setup, based on it being the version available for the
"mariadb-server" package in Debian stable.

The Docker container is configured to create the "lobsters_dev" database
automatically, so it shouldn't be necessary to run a command to explicitly
create the database anymore.

Fixes #1050.