fix voting/flagging from a7299e84
If the Lobster replacement class isn't in the global scope, the existing
Lobsters can't refer to it. But if it's in the global scope, the DOM isn't yet
loaded for it to read the current username. When the old class is entirely
replaced, insantiation of the class can move back into onPageLoad().