don't prevent user deletion if they have a message with a doffed hat
Also, this removes message undeletion on user reactivation. That behavior is
really surprising, doesn't discriminate between messages the user individually
deleted versus bulk-deleted when they closed their account, and only works on
messages the counterpart hasn't also deleted. That's a lot of odd corner cases
that are not explained to the user and in any case it is probably negative value
to undelete messages when doing a different task.

Fix #1267