Dump heap after boot and after bloat
Note: This is totally untested

`out_of_band` is invoke when a request finishes and no other
thread is busy. So theorically it may never be invoked if the
worker is under constant load, but I'm hoping it will eventually.

The idea here is to wait for the number of live object to go over
some threshold we found in the GC stqats we dumped previously, and
when it's reached we dump the heap a second time.

By comparing it with the heap that was dumped after boot, we should
be able to figure out what extra objects are being retained after
the worker ages.