Precompute Comment scores on initial insertion
I was looking at https://github.com/lobsters/lobsters/issues/1238
and the comment insertion transaction is indeed very meaty, over 22
queries.

Looking into how to merge or eliminate some of them, I noticed
that the comment is quickly updated right after insertion, from
`update_score_and_recalculate!`, and that it's quite a big query
for something that always amount to the same result:

  - A new comment only has the author's vote.
  - It also has no flags.
  - And the confiderence is directly derived from the votes.

As such we can skip all this and just insert default values.

This reduce the number of queries in that transaction from 22 to 16.

The eliminated queries however are likely not the one causing contention,
It's way more likely to be the one that update the story, but it may
still help a bit by making the transaction shorter.