Avoid continiously instantiating HTMLEntities
While profiling the lobsters benchmark in the yjit-bench suite
I noticed a weird 0.8% of time spent in `instance_eval`.

Tracking it down, it's coming from the htmlentities gem,
when calling `HTMLEntities.new` it generates code on the fly
based on the options.

As such you aren't meant to instantiate a new one every time, but
have a single global object somewhere.

Otherwise this constant code generation is wasting performance
for now reason, and potentially use YJIT executable memory
if called enough.