Broad strokes port to Action Mailbox
Some finer points to clean up (and tests to write), but functionality
for replying to stories and comments ported from old scripts to Action
Mailbox. It's been tested as much as I can with my understanding of
the feature and RoR.

I'm unsure whether Active Storage *is* something we can enable, but I
haven't found a way to disable it with Action Mailbox so keeping it
around for now.