Upgrade to Rails 8.0
Beside what the number may suggest, for as much as lobsters is
concerned, this is a very minor upgrade. Most of Rails 8.0 is just
Rails 7.2 with a few new defaults.

The only small issue caused by the upgrade was that `has_secure_password`
now adds a methods to generate a password reset token that don't
need to be perisisted, and that method did happen to conflict with
the existing `password_reset_token` attribute.