Remove useless private Cop
This cop never did anything because `standardrb` isn't configured to
load it, and also because it's not capable of parsing ERB files anyway.

To do that you'd need to bring `erb-lint` or something like that.

I have branch that tries to make it work, but it's a big hassle,
so in my humble opinion it's not worth it.