Update URL when fetching title
When the user clicks 'Fetch Title', in addition to filling in the title field,
also automatically set the value of the URL field. This is primarily because
servers often redirect to a canonicalized or secure (HTTPS) version of a URL if
someone requests an alternate or insecure (HTTP) version.

With this change, we use the URL that the server prefers for the document.