add overview suggestion to brigading guideline
Linking into other community's spaces isn't always bad, but it is the core of
brigading, so it's worth avoiding on contentious topics. We don't want to dump
all of Lobsters traffic into some unexpecting project's already difficult github
issue or start out our by misunderstanding a hard topic because we are reading
an email assuming it would only be read by a small group with a lot of context.