simplify find_or_create_origin from url by using find_or_create_by!
find_or_create_by! will attempt to create the record if it doesn't
already exist, but if that fails with `ActiveRecord::RecordNotUnique`,
it will assume it encountered a race condition and will try to find the
record once more.
https://api.rubyonrails.org/v8.0.1/classes/ActiveRecord/Relation.html#method-i-find_or_create_by

I deleted the `indentifier=` setter in the Origin model, because if
origin is created from the Domain model, it should already have the link
to the domain record.