Only use params we expect in moderation index
These are used to generate the pagination links, which will break if `:controller` or `:action` are passed as parameters due to Rails internals. We shouldn't be "forwarding" random parameters from the user, but generating a link only with the params we expect to be in the URL.

This doesn't currently appear to be a security issue, but it would've been a useful place for injecting things into links on the page if required to pair with another attack later.