Proxy Scenic to either MySQL or SQLite adapters
This allows us to manage views in both databases, albeit with a layer of indirection added into the stack. We want to run the primary database on MySQL but the (solid)cache database on SQLite.

Without doing this we either can't manage the primary or the cache database, especially running `bin/rails db:schema:dump`.