link to office hours stream on /about and when live
The schedule is pretty stable at this point. I've been trying not to use
office hours as self-promo (see 2b44cf6), but some recent viewers and chatters
have been very surprised to find it exists now that the announcement post is
long gone from the front page.

If you don't want to see it, the adblock rule is: lobste.rs###office-hours-live