block common brigading targets
Because I know this strict enforcement of brigading is only ~4 years old: the
big reason is scale. Lobsters is huge now. We can dump 100k logged-in users into
someone's random fight in a github issue. Think of how disruptive that would be
to one of your own repos. We're tiny by the scale of HN or Reddit, but huge for
a disagreement about licensing or UI change in some random dev's 10kloc hobby
project.