Lay out the pages for private messages using CSS Grid
This was pretty smooth, requiring only a few changes to the stylesheet
itself. Hopefully as this work goes on it'll just be a matter of putting
the appropriate classes in the appropriate places.

The table for the list of messages is still laid out with manual
percentage widths. This might be worth changing at some point, but for
now I'm aiming to standardise the common page structures, rather than
the unique elements.