Add a CSP header in report-only mode
This is a tighter policy than the one commented out before. Due to the
JavaScript importmap counting as an inline script, we need to use nonces
to allow that without also allowing all inline scripts.

This switches to a random nonce rather than deriving it from the
session. The benefit of the session ID approach is to enable conditional
caching, but the site does not make use of this. Using a random nonce
reduces the chance of security issues (for example, logging in currently
doesn't reset the session as far as I can see) and means we can generate
nonces for guest requests who hit Rails.

For guest users who hit the cache, we need to set a CSP in Caddy as
well. As we can't use nonces here, I've opted to disallow scripts
altogether. This approach fails an edge case on the signup page, where a
guest user is served JavaScript for the markdown formatting dropdown. In
this situation, that script and functionality won't work.

For now this is all using report-only mode so that any oversights don't
break the site for users.