Reinstate `/top/rss` and allow RSS for any `/top/…/rss` path (#1629)
* Consistently route /top/rss with all other params

Also add a spec to make sure they stay routeable

* Resurrect /top/rss feed and implement for all top paths

We were redirecting /top/rss previously without maintaining the format so it wasn't returning an RSS feed at any point. Add some tests to cover that and fix it so we redirect with the length maintained now.

Allow it to return RSS for any of the /top/…/rss paths, because its not hard and someone might want it