Add inbox_mentions setting
Since people keep mentioning cadey in the friendlysock thread, this
seems like a good time to add a setting.

This is a really basic subset of https://github.com/lobsters/lobsters/issues/872

I can add the other booleans as well, but thought i'd start with the most minimal needed change.