Add handling for active_stories_count
Render 404, if there are no "active" stories to be shown on the /domains/X page.
For a user that is not a moderator, the number of "active" stories for "active" submitters will be shown.
For a user that is a moderator, the total number of stories from all submitters will be shown.
Note: for a user that is not a moderator, who deleted their own story,
they will see the number of "active" stories but the list of stories will include all their deleted story.