Add checksums to Gemfile.lock for bundler
Generated by running

    $ bundle lock --add-checksums

These are updated whenever bundler adds or updates a gem in our dependencies, and ensures if a gem changes upstream after we've initially pinned it then we can detect it. (eg, someone changes the .gem file for a given version of a gem.) It's a mitigation against a supply chain attack.