Upgrade Bundler, add more platforms to lockfile & checksum dependencies for security (#1747)
* Upgrade to Bundler 2.7.2

* Add ruby platform to Gemfile.lock as fallback

If gems don't exist as precompiled versions for linux and/or mac then allow bundler to compile them instead.

* Add macOS to bundler platforms in lockfile

Without this it constantly gets added anytime you run bundler on a mac, won't affect production but means developers on macs don't have issues.

* Add checksums to Gemfile.lock for bundler

Generated by running

    $ bundle lock --add-checksums

These are updated whenever bundler adds or updates a gem in our dependencies, and ensures if a gem changes upstream after we've initially pinned it then we can detect it. (eg, someone changes the .gem file for a given version of a gem.) It's a mitigation against a supply chain attack.