Add dependabot.yml that checks only for brakeman (#1786)
* Add dependabot.yml that checks only for brakeman

* Remove prism from Gemfile

738069c added Prism as a direct dependency in order to work around
Brakeman issue #1927.

More recently, 236bce1 made Prism 1.4 an indirect dependency, so Prism
longer needs to be listed in the Gemfile.

* Remove --ensure-latest option from Brakeman CI step

No need for this option now that we're using Dependabot to keep Brakeman
up to date (9c809e2c). Removing it means a better contributor
experience, as PRs won't be blocked by CI failures from this.