Telebugs: fix user enrichment
prepare_exception_notifier was getting called after the
authenticate_user injected by the Authenticatable Concern, so its blank
info was overwriting the enriched user info. In any case, we want to
prepare as early as possible so we have a partially-enriched report if
one of the other before_actions or framework layers throws.