Fix database_consistency MissingIndexChecker, ForeignKeyChecker
From the database_consistency gem documentation:

MissingIndexChecker:
https://github.com/djezzzl/database_consistency/wiki/missingindexchecker

Imagine your model has a `has_one :user` association but has no index in
the database. In this case, querying the database to get the associated
instance can be very inefficient. Mostly, you'll need an index to
process such queries fast.

Additionally, `has_one` associations should be unique on the database
level to avoid unexpected behavior.

ForeignKeyChecker:
https://github.com/djezzzl/database_consistency/wiki/foreignkeychecker

Imagine your model has `belongs_to :user`. It can happen that the user,
it's being belonging to, may not be existing anymore in the database.
This could bring bugs and in order to ensure data consistency, you need
to have a foreign key constraints in the database.